Make SmallDrinkCell reusable for see more.
