--- CocktailRecipes/CocktailRecipes/Features/Home/Charts/Views/DrinkChartCell.swift
@@ -1,80 +1,0 @@
-//
-//  DrinkChartCell.swift
-//  CocktailRecipes
-//
-//  Created by Jaison Vieira on 10/17/20.
-//  Copyright © 2020 Eric Rado. All rights reserved.
-//
-
-import UIKit
-
-final class DrinkChartCell: UICollectionViewCell {
-    private let imageView: UIImageView = {
-        let imageView = UIImageView()
-        imageView.backgroundColor = .gray
-        imageView.translatesAutoresizingMaskIntoConstraints = false
-        imageView.contentMode = .scaleAspectFit
-        return imageView
-    }()
-
-    private let rankLabel: UILabel = {
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.font = UIFont.preferredFont(forTextStyle: .body)
-        label.contentHuggingPriority(for: .vertical)
-        return label
-    }()
-
-    private let titleLabel: UILabel = {
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.font = UIFont.preferredFont(forTextStyle: .body)
-        label.contentHuggingPriority(for: .vertical)
-        return label
-    }()
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        setupView()
-    }
-
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    override func prepareForReuse() {
-        super.prepareForReuse()
-        imageView.image = nil
-        titleLabel.text = nil
-    }
-
-    private func setupView() {
-        contentView.addSubview(imageView)
-        contentView.addSubview(rankLabel)
-        contentView.addSubview(titleLabel)
-
-        NSLayoutConstraint.activate([
-            imageView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 16),
-            imageView.topAnchor.constraint(equalTo: contentView.topAnchor, constant: 8),
-            imageView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor, constant: -8),
-            imageView.widthAnchor.constraint(equalTo: imageView.heightAnchor)
-        ])
-
-        NSLayoutConstraint.activate([
-            rankLabel.leadingAnchor.constraint(equalTo: imageView.trailingAnchor, constant: 24),
-            rankLabel.topAnchor.constraint(equalTo: contentView.topAnchor, constant: 10)
-        ])
-
-        NSLayoutConstraint.activate([
-            titleLabel.leadingAnchor.constraint(equalTo: imageView.trailingAnchor, constant: 24),
-            titleLabel.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -16),
-            titleLabel.centerYAnchor.constraint(equalTo: contentView.centerYAnchor)
-        ])
-    }
-
-    func configure(image: UIImage?, rank: Int, text: String) {
-        imageView.image = image
-        rankLabel.text = "#\(rank)"
-        titleLabel.text = text
-    }
-}
