+++ CocktailRecipes/CocktailRecipes/Features/Home/Views/SmallDrinkCell.swift
@@ -9,57 +9,79 @@
 import UIKit
 
 final class SmallDrinkCell: UICollectionViewCell {
-	private let imageView: UIImageView = {
-		let imageView = UIImageView()
-		imageView.backgroundColor = .gray
-		imageView.translatesAutoresizingMaskIntoConstraints = false
-		imageView.contentMode = .scaleAspectFit
-		return imageView
-	}()
+    private let imageView: UIImageView = {
+        let imageView = UIImageView()
+        imageView.backgroundColor = .gray
+        imageView.translatesAutoresizingMaskIntoConstraints = false
+        imageView.contentMode = .scaleAspectFit
+        return imageView
+    }()
 
-	private let label: UILabel = {
-		let label = UILabel()
-		label.translatesAutoresizingMaskIntoConstraints = false
-		label.font = UIFont.preferredFont(forTextStyle: .body)
-		label.contentHuggingPriority(for: .vertical)
-		return label
-	}()
+    private let rankLabel: UILabel = {
+        let label = UILabel()
+        label.translatesAutoresizingMaskIntoConstraints = false
+        label.font = UIFont.preferredFont(forTextStyle: .body)
+        label.contentHuggingPriority(for: .vertical)
+        return label
+    }()
 
-	override init(frame: CGRect) {
-		super.init(frame: frame)
-		setupView()
+    private let titleLabel: UILabel = {
+        let label = UILabel()
+        label.translatesAutoresizingMaskIntoConstraints = false
+        label.font = UIFont.preferredFont(forTextStyle: .body)
+        label.contentHuggingPriority(for: .vertical)
+        return label
+    }()
+
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        setupView()
+    }
+
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    override func prepareForReuse() {
+        super.prepareForReuse()
+        imageView.image = nil
+        rankLabel.text = nil
+        titleLabel.text = nil
+    }
+
+    private func setupView() {
+        contentView.addSubview(imageView)
+        contentView.addSubview(rankLabel)
+        contentView.addSubview(titleLabel)
+
+        NSLayoutConstraint.activate([
+            imageView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 16),
+            imageView.topAnchor.constraint(equalTo: contentView.topAnchor, constant: 8),
+            imageView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor, constant: -8),
+            imageView.widthAnchor.constraint(equalTo: imageView.heightAnchor)
+        ])
+
+        NSLayoutConstraint.activate([
+            rankLabel.leadingAnchor.constraint(equalTo: imageView.trailingAnchor, constant: 24),
+            rankLabel.topAnchor.constraint(equalTo: contentView.topAnchor, constant: 10)
+        ])
+
+        NSLayoutConstraint.activate([
+            titleLabel.leadingAnchor.constraint(equalTo: imageView.trailingAnchor, constant: 24),
+            titleLabel.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -16),
+            titleLabel.centerYAnchor.constraint(equalTo: contentView.centerYAnchor)
+        ])
 	}
 
-	required init?(coder: NSCoder) {
-		fatalError("init(coder:) has not been implemented")
-	}
-
-	override func prepareForReuse() {
-		super.prepareForReuse()
-		imageView.image = nil
-		label.text = nil
-	}
-
-	private func setupView() {
-		contentView.addSubview(imageView)
-		contentView.addSubview(label)
-
-		NSLayoutConstraint.activate([
-			imageView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 16),
-			imageView.topAnchor.constraint(equalTo: contentView.topAnchor, constant: 8),
-			imageView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor, constant: -8),
-			imageView.widthAnchor.constraint(equalTo: imageView.heightAnchor)
-		])
-
-		NSLayoutConstraint.activate([
-			label.leadingAnchor.constraint(equalTo: imageView.trailingAnchor, constant: 24),
-			label.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -16),
-			label.centerYAnchor.constraint(equalTo: contentView.centerYAnchor)
-		])
-	}
-
-	func configure(image: UIImage?, text: String) {
+    func configure(image: UIImage?, text: String, rank: Int?) {
 		imageView.image = image
-		label.text = text
+        titleLabel.text = text
+
+        if let rank = rank {
+            rankLabel.text = "\(rank)"
+            rankLabel.isHidden = false
+        } else {
+            rankLabel.isHidden = true
+        }
 	}
 }
